Added the controller and bot interface
The Controller Class abstracts the IO to the AEI controller and
interfaces with the bot that makes all the game playing decisions.

The Bot Abstract Class shows the interface that the Controller
Class uses. Bot implementations should just inherit from Bot and
override all the virtual functions.

Currently, the setPosition and setOption commands are not
supported.